requirements.txt
Authlib==0.15.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
Flask==1.1.2
h11==0.12.0
httpcore==0.12.3
httpx==0.17.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
oauthlib==3.1.0
prompt-toolkit==3.0.18
pycparser==2.20
PySocks==1.7.1
python-dateutil==2.8.1
regex==2021.3.17
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.4.0
selenium==3.141.0
six==1.15.0
sniffio==1.2.0
tda-api==1.1.8
tqdm==4.59.0
urllib3==1.26.4
wcwidth==0.2.5
websockets==8.1
Werkzeug==1.0.1
